Run tests over the official datasets
I've added a `golden.csv` file which contains the first 50 records from
the January 2018 Ford GoBike trip data set.

If any of the CSVs from the official dataset are placed into the
testdata directory, the test suite will automatically discover and run
the tests and benchmarks using those files in addition to the golden
set.