Add data about station capacity
Also add data about weekday trips per dock, which will help give
a better sense of the utilization of a station, along with data about
stations that are consistently empty. What we want to capture is the
number of people that would have rented a bike if there were more
bikes, or the number of people that would have parked a bike if there
was more room, but that's difficult, so we have to approximate based
on usage data.

Rework the client to add caching support, and allow retrieving
stations as well. Add stations as a Make target so you can have local
fallback if you're not on the internet.